vita-django/requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.53
botocore==1.17.63
bcrypt==3.1.7
blinker==1.4
braintree==4.4.0
Brlapi==0.7.0
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.2
django-storages==1.10
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-social-oauth2==1.0.8
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
ipaddr==2.2.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pep517==0.8.2
pexpect==4.6.0
Pillow==7.2.0
pipenv==11.9.0
progress==1.5
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
systemd-python==234
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
